Increase test coverage, fix status
Increase the test coverage for the charm class and fix the status
reporting so that the user can see errors occuring. This change
introduces custom status checks and moves the service running status
check to use the RESTART_MAP, which is a mapping of configuration files
to services that should be restarted as a result of the configuration
file changing on disk.

The first custom status check introduced will determine if the debian
package provided as a resource has been provided or not. If the user
has not provided the package, the unit status will report that it is
Blocked waiting for the user to provide the debian package.

The second custom status check introduced will determine if the
installation was successful. This is a simple check relying on stored
state. This works for now but can be improved by checking the debian
package install status using dpkg -l <package-name> and parsing the
output. A simple approach was chosen for the time being.

The install_sdc method was refactored to leverage the common code used
by the new status checks, for consistency. A new MaintenanceStatus was
added for the unit prior to installing the SDC kernel module in order
to indicate that the kernel module is being installed.

Signed-off-by: Billy Olsen <billy.olsen@canonical.com>